\name{GAabbreviation-internal}
\alias{GAabbreviation-internal}
\alias{alphas.GAabbreviate}
\alias{convCorr.GAabbreviate}
\alias{fitness.GAabbreviate}
\alias{GAabbreviation-package}
\alias{is.missing}
\alias{impute}
\alias{makeKey.GAabbreviate}
\alias{measure.GAabbreviate}
\alias{monitor.GAabbreviate}
\alias{plot.GAabbreviate}
\alias{print.GAabbreviate}
\alias{summary.GAabbreviate}

\title{Internal GAabbreviation functions}

\description{Internal functions not intended to be called directly by users or functions for which no documentation has bee written so far.}

%\usage{}

\keyword{internal}